added many things
added timer, level menu button methods, fixed bug with level generation, added multiple pages to level selector and corresponding scripts, adjusted ui placement for game board,

timer starts when any key is pressed and ends when the game ends. not yet any storage for times though...

level select menu has script that takes in array of "pages", which have been saved as a prefab and contain three panels, each with an icon, some text and other stuff, and a button function that has to be assigned.

the gameboard's ui needs to be adjusted to be less cancer -- we should do that at the next meeting

there was a bug with generating levels and going back and forth between levels. one was that it wasn't loading the data right and the other was that when reloading the level selector screen there would be another copy of the intermediary object, since it is set to do not destroy.