editted level gen script and added end game screen
the level gen script was getting messy so i separated it into level gen, which operates solely for the generation of the level/adjustment of camera

and playercontrols, which controls the players and also the text. if this file gets too messy I may separate it further but that would be... rather difficult.

i also made an end screen for when the player reaches the end of the game board, which includes something along the lines of "congrats, your time is 00:10" or something, with buttons to go to the next level, retry the current level, go to level select, or quit.

while this end screen is up, the escape function for pause menu should not work.

the time displays correctly and the buttons function correctly, but as of right now the times are not saved. a next step would be to learn how to save data and display it on the level select panels